:bug: Take care of singularities: DCM => Euler angles
There are some cases in which the conversion between DCM and Euler
angles were not correct. For example, if a DCM was a rotation about
X-axis, then the conversion to Euler angles using the rotation
sequence `:XYX` or `:XZX` would provide an incorrect result.

This was also affecting the conversion from Quaternions to Euler
angles because `quat_to_angle` uses `dcm_to_angle`.