:warning: Drop supertype of Quaternion
Previously, Quaternion was `<:AbstractVector`. However, this choice was
leading to many problems when interfacing with other packages. For
example, it was very difficult to make it works together with Zygote.jl
because of the multiplication. In the previous version, Quaternion was
an array in which the multiplication `q1 * q2` (both 4x1 arrays) leads
to another 4x1 arrays, breaking a lot of assumptions about arrays.

Many functions were defined to reduce the number of breakage. Quaternion
supports iterations and broadcast. Hence, I do not expect many problems.